Cleaned up source code and formatted it

+++ src/main/java/com/DavidM1A2/AfraidOfTheDark/client/MCAClientLibrary/MCAModelBox.java
@@ -21,9 +21,9 @@ public class MCAModelBox extends ModelBox
 
 		this.MCAquadList = new TexturedQuad[6];
 
-		float endX = posX + (float) sizeX;
-		float endY = posY + (float) sizeY;
-		float endZ = posZ + (float) sizeZ;
+		float endX = posX + sizeX;
+		float endY = posY + sizeY;
+		float endZ = posZ + sizeZ;
 		posX -= scaleFactor;
 		posY -= scaleFactor;
 		posZ -= scaleFactor;
