+++ src/main/java/com/DavidM1A2/AfraidOfTheDark/common/dimension/nightmare/NightmareTeleporter.java
@@ -37,7 +37,7 @@ public class NightmareTeleporter extends Teleporter
 			int i = MathHelper.floor_double(entity.posX);
 			int j = MathHelper.floor_double(entity.posY) - 1;
 			int k = MathHelper.floor_double(entity.posZ);
-			entity.setLocationAndAngles((double) i, (double) j, (double) k, entity.rotationYaw, 0.0F);
+			entity.setLocationAndAngles(i, j, k, entity.rotationYaw, 0.0F);
 			entity.motionX = entity.motionY = entity.motionZ = 0.0D;
 		}
 		else if (dimensionOld == Constants.NightmareWorld.NIGHTMARE_WORLD_ID)
